Fix exception thrown in service_delivery_callback_task
The odd time we get a SES callback for a service with a short data retention (1 day), since there isn't a to field in the notification sent to create the status callback an exception is thrown. This PR fixes that by checking if the data type of the object passed in.

Question: what should the "notification_to" property be. I don't like returning None here as that might cause an exception with the service. Perhaps "recipient not available" but what if an invalid email address causes an issue as well.