🔨 Add script to set sensible DB defaults in dev
When the application is first setup using `scripts/bootstrap.sh`, the
database is seeded with some default records that aren't entirely suited
to local development. For example, the default SMS sender is set to
"GOVUK" (a sender ID that'll get you banned with SMS backends) as well
as seeding a default user without any permissions and a randomly
generated (and therefore unknown) login password.

This commit adds a new `apply_dev_defaults.py` script, adding/updating
database records so it's easier for a developer to login and get up and
running with minimal fuss.