Remove redundant override for receipts worker
For some reason this doesn't work in practice: STATSD host is set
to the expected host when I run `cf env ...` in all environments.
Conversely, it is set to "None" for notify-api. I suspect this has
something to do with specifying the value twice [1][2].

It's unclear why this was disabled when the original work (renamed
from callbacks [3]) was added [4]. Enabling StatsD means we'll get
metrics for tasks executed by this worker, which we're missing.

(I suspect the config was copied from the app manifest [5], where it
makes sense, since the app can expose a /metrics endpoint instead.)

[1]: https://github.com/alphagov/notifications-api/blob/f3fdd3b09b15f9ce6787073f892d3a38df5eee14/manifest.yml.j2#L132
[2]: https://github.com/alphagov/notifications-api/blob/f3fdd3b09b15f9ce6787073f892d3a38df5eee14/manifest.yml.j2#L153
[3]: https://github.com/alphagov/notifications-api/pull/3206
[4]: https://github.com/alphagov/notifications-api/commit/9d5b629dadb2ec057569f77414a7d131bbb5035e
[5]: https://github.com/alphagov/notifications-api/blob/39d7c347b047966bb56b1805ced0463e36e6654c/manifest.yml.j2#L7