Warn about SMS failures when it's cost effective
Previously we warned when as little as 100 SMS messages failed, which
equates to £1.60 in cost. This increases the absolute threshold so that
we avoid taking action when the cost in developer time is greater than
the money wasted. If it takes us an hour to solve the ticket, then that
equates to roughly £100 in people costs.

In future, we should also consider simplifying this check to be about
permanent failures in general.