Move request_id injection into send_task override
This applies the same change we made in other apps [1][2]. Adding
the override here is special, though, because it means the others
will now get triggered, since this app is the start of the chain
of tasks for a request. We will also retain existing request_id
tracing for tasks within this app, since "apply_async" calls the
"send_task" method internally, which is the one we're overriding.

[1]: https://github.com/alphagov/notifications-template-preview/pull/531/commits/6f3c118a1e071f8169698806450b45aa685859aa
[2]: https://github.com/alphagov/notifications-antivirus/pull/69/commits/2e08b7aa954c120eb0265fda41ad2025616426ba